Feat: create database creation pipeline
