+++ app/database/enums.py
@@ -1,8 +1,4 @@
-from typing import Dict, List
-from pydantic import BaseModel, ConfigDict
 from sqlmodel import Enum
-
-from app.database.models import Subject
 
 
 class SubjectClassStatus(str, Enum):
@@ -70,33 +66,3 @@ class ChunkType(str, Enum):
     table = "table"
     other = "other"
     # NOTE: add more types as needed, but keep clean structure with good segregation
-
-
-# NOTE: not really an enum, but fits here for now
-class ClassInfo(BaseModel):
-    """Maps subjects to their grade levels for a teacher"""
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    subjects: Dict[str, List[str]]  # keys=Subject, values=List[GradeLevel]
-
-    """ METHODS """
-
-    def model_dump(self, **kwargs):
-        data = super().model_dump(**kwargs)
-        return {
-            subject: [grade for grade in grades]
-            for subject, grades in data["subjects"].items()
-        }
-
-    # TODO: Double check this validator
-    @classmethod
-    def model_validate(cls, data: Dict):
-        if data is None:
-            return None
-        return cls(
-            subjects={
-                Subject(subject): [GradeLevel(grade) for grade in grades]
-                for subject, grades in data.items()
-            }
-        )
