+++ app/tools/tool_code/generate_exercise/main.py
@@ -4,10 +4,11 @@ from typing import List, Optional
 from app.utils.llm_utils import async_llm_request
 from app.utils.prompt_manager import prompt_manager
 from app.database.db import vector_search
-from app.database.models import Chunk, ChunkType, Resource, User
+from app.database.models import Chunk, Resource, User
 from app.config import llm_settings
 from app.services.whatsapp_service import whatsapp_client
 from app.utils.string_manager import strings, StringCategory
+from app.database.enums import ChunkType
 
 logger = logging.getLogger(__name__)
