+++ scripts/assets/sample_data/data.yaml
@@ -1,4 +1,4 @@
-geography_textbook:
+geography_resource:
   name: "Geography for Secondary Schools Student's Book Form Two"
   type: "textbook"
   authors:
